+++ R/get.R
@@ -482,13 +482,8 @@ get_usage_static <- function(src, content_guid = NULL,
 get_usage <- function(client, from = NULL, to = NULL) {
   from <- format(from, "%Y-%m-%dT%H:%M:%SZ")
   to <- format(to, "%Y-%m-%dT%H:%M:%SZ")
-  usage_raw <- client$GET(
-    connectapi:::v1_url("instrumentation", "content", "hits"),
-    query = list(
-      from = from,
-      to = to
-    )
-  )
+
+  usage_raw <- client$inst_content_hits(from, to)
 
   parse_connectapi_typed(usage_raw, connectapi_ptypes$usage)
 }